+++ src/components/test.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const display = "Hello, to Guru99 Tutorials";
+const h1tag =<h1>{display}</h1>;
+export default h1tag;