+++ src/containers/Layout.js
@@ -1,0 +1,15 @@
+import React from 'react'; 
+   
+class ExpenseEntryItem extends React.Component {
+  render() {
+     return (
+        <div>
+           <div><b>Item:</b> <em>Mango Juice</em></div>
+           <div><b>Amount:</b> <em>30.00</em></div>
+           <div><b>Spend Date:</b> <em>2020-10-10</em></div>
+           <div><b>Category:</b> <em>Food</em></div>
+        </div>
+     );
+  }
+}
+export default ExpenseEntryItem;